backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
shap==0.43.0
google-generativeai==0.3.1
python-dotenv==1.0.0
